Add pronunciation scraper script
It is unclear how this script exactly works with respect to the web
interface it scrapes from. On some platforms I've tested it on, it works
fine, whereas on others, it does not. Since it works provisionally on my
machine for now, that will have to do, for the temporary purposes of
scraping the desired data, anyway.